requirements.txt
python-telegram-bot
llama-cpp-python